Updated README.md
I revised the summary of your project a bit and added some extra information that might be useful to incoming users. I added a table of content, installation section for adding the bot to a discord server, features and commands section for all of the bot's commands, and a way for people to contact you for any questions.